requirements.txt
backports.csv==1.0.7
beautifulsoup4==4.9.0
certifi==2020.4.5.1
chardet==3.0.4
contextlib2==0.6.0.post1
docopt==0.6.2
ffmpeg-python==0.2.0
future==0.18.2
idna==2.8
internetarchive==1.9.3
jsonpatch==1.25
jsonpointer==2.0
mutagen==1.44.0
requests==2.22.0
schema==0.7.2
six==1.14.0
soupsieve==2.0
total-ordering==0.1.0
tqdm==4.46.0
urllib3==1.25.9
